+++ local-store/LocalStoreByTimestamp.js
@@ -1,8 +1,9 @@
 "use strict";
 var ClearFullStore = require("./ClearFullStore");
 var UniqueChronologicalKeys = require("./UniqueChronologicalKeys");
-/** LocalStoreByTimestamp namespace
- * persistent storage interface for small objects or data blobs
+/** LocalStoreByTimestamp is a UniqueStore implementation whick takes a 'keyGenerator'.
+ * The key generator creates keys for the addItem() method instead of the caller providing a key as with LocalStore.setItem()
+ * @see UniqueStore
  * @see LocalStore
  * @author TeamworkGuy2
  * @since 2016-3-25